requirements.txt
customtkinter
requests
python-docx
PyPDF2
foundry-local
